Updated part 3 and final part 4
part 4 uses the function from part 3 to call data.
The project told you not to do it, but I'm also in the business where you should take advantage of the work you just did to save time. I'll go back and make a standalone function for part 4.

Note the last 3 chunks in scrap work are how they wanted to call data. I find it ridiculous that they wanted you to find the hospital which is just as easy to find using P3 f(). But, I guess I should still do it for the practice no matter how silly it is,